add CloseChannel refactor, separated in to smaller components.
